+++ src/lifecycle/run.rs
@@ -154,17 +154,15 @@ fn run_impl<T: State, F: FnOnce()->Result<T>>(title: &str, size: Vector, setting
 
     let key_names = generate_key_names();
     handle_event(&document, &app, move |mut app, event: KeyDownEvent| {
-        // If the key press is 'Backspace', we shouldn't send a typed event
-        // However, if it is a single alphanumeric character, that should be a typed event
-        // TODO: this is imperfect at best, a better way to decide what codes get sent to the
-        // application would be desirable
-        let string = event.key();
-        let mut characters = string.chars();
+        // Winit doesn't filter to printable Typed events, so it should just be up to the user
+        // However there are control keys (think Backspace, Shift, Super) so only report a key code
+        // if it is a single character
+        let key = event.key();
+        let mut characters = key.chars();
         let first = characters.next();
         let second = characters.next();
-        match (first, second) {
-            (Some(ch), None) if ch.is_alphanumeric() => app.event_buffer.push(Event::Typed(ch)),
-            _ => ()
+        if let (Some(key), None) = (first, second) {
+            app.event_buffer.push(Event::Typed(key));
         }
         if let Some(keycode) = key_names.get(&event.code()) {
             app.event_buffer.push(Event::Key(KEY_LIST[*keycode], ButtonState::Pressed));
@@ -329,7 +327,7 @@ static KEY_NAMES: &[&str] = &[
     "Numpad9",
     "AbntC1",
     "AbntC2",
-    "Add",
+    "NumpadAdd",
     "Quote",
     "Apps",
     "At",
@@ -337,11 +335,11 @@ static KEY_NAMES: &[&str] = &[
     "Backslash",
     "Calculator",
     "Capital",
-    "Colon",
+    "Semicolon",
     "Comma",
     "Convert",
-    "Decimal",
-    "Divide",
+    "NumpadDecimal",
+    "NumpadDivide",
     "Equal",
     "Backquote",
     "Kana",
@@ -355,7 +353,7 @@ static KEY_NAMES: &[&str] = &[
     "MediaSelect",
     "MediaStop",
     "Minus",
-    "Multiply",
+    "NumpadMultiply",
     "Mute",
     "LaunchMyComputer",
     "NavigateForward",
@@ -379,7 +377,7 @@ static KEY_NAMES: &[&str] = &[
     "Slash",
     "Sleep",
     "Stop",
-    "Subtract",
+    "NumpadSubtract",
     "Sysrq",
     "Tab",
     "Underline",
@@ -509,7 +507,7 @@ mod tests {
         assert_eq!(Key::Numpad9 as usize, mapping["Numpad9"]);
         assert_eq!(Key::AbntC1 as usize, mapping["AbntC1"]);
         assert_eq!(Key::AbntC2 as usize, mapping["AbntC2"]);
-        assert_eq!(Key::Add as usize, mapping["Add"]);
+        assert_eq!(Key::Add as usize, mapping["NumpadAdd"]);
         assert_eq!(Key::Apostrophe as usize, mapping["Quote"]);
         assert_eq!(Key::Apps as usize, mapping["Apps"]);
         assert_eq!(Key::At as usize, mapping["At"]);
@@ -517,11 +515,11 @@ mod tests {
         assert_eq!(Key::Backslash as usize, mapping["Backslash"]);
         assert_eq!(Key::Calculator as usize, mapping["Calculator"]);
         assert_eq!(Key::Capital as usize, mapping["Capital"]);
-        assert_eq!(Key::Colon as usize, mapping["Colon"]);
+        //assert_eq!(Key::Colon as usize, mapping["Semicolon"]);
         assert_eq!(Key::Comma as usize, mapping["Comma"]);
         assert_eq!(Key::Convert as usize, mapping["Convert"]);
-        assert_eq!(Key::Decimal as usize, mapping["Decimal"]);
-        assert_eq!(Key::Divide as usize, mapping["Divide"]);
+        assert_eq!(Key::Decimal as usize, mapping["NumpadDecimal"]);
+        assert_eq!(Key::Divide as usize, mapping["NumpadDivide"]);
         assert_eq!(Key::Equals as usize, mapping["Equal"]);
         assert_eq!(Key::Grave as usize, mapping["Backquote"]);
         assert_eq!(Key::Kana as usize, mapping["Kana"]);
@@ -535,7 +533,7 @@ mod tests {
         assert_eq!(Key::MediaSelect as usize, mapping["MediaSelect"]);
         assert_eq!(Key::MediaStop as usize, mapping["MediaStop"]);
         assert_eq!(Key::Minus as usize, mapping["Minus"]);
-        assert_eq!(Key::Multiply as usize, mapping["Multiply"]);
+        assert_eq!(Key::Multiply as usize, mapping["NumpadMultiply"]);
         assert_eq!(Key::Mute as usize, mapping["Mute"]);
         assert_eq!(Key::MyComputer as usize, mapping["LaunchMyComputer"]);
         assert_eq!(Key::NavigateForward as usize, mapping["NavigateForward"]);
@@ -559,7 +557,7 @@ mod tests {
         assert_eq!(Key::Slash as usize, mapping["Slash"]);
         assert_eq!(Key::Sleep as usize, mapping["Sleep"]);
         assert_eq!(Key::Stop as usize, mapping["Stop"]);
-        assert_eq!(Key::Subtract as usize, mapping["Subtract"]);
+        assert_eq!(Key::Subtract as usize, mapping["NumpadSubtract"]);
         assert_eq!(Key::Sysrq as usize, mapping["Sysrq"]);
         assert_eq!(Key::Tab as usize, mapping["Tab"]);
         assert_eq!(Key::Underline as usize, mapping["Underline"]);